services/risk-engine-qml/requirements.txt
# --- Quantum Core (Qiskit 2.x Ecosystem) ---
qiskit==2.2.0
qiskit-aer==0.16.0          # High-performance classical simulator
qiskit-ibm-runtime==0.36.0  # Required for IBM QPU execution
qiskit-machine-learning==0.9.0  # Migrated to V2 primitives for 2.x compatibility

# --- Machine Learning & Data ---
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0         # For classical pre-processing (StandardScaler)

# --- Infrastructure & Messaging ---
nats-py==2.7.2              # To listen for "Governor" instructions
redis==7.4.0                # To store circuit results or state
prometheus-client==0.20.0   # To export "Circuit Execution Time" metrics

# --- Performance ---
pydantic==2.6.1             # For payload validation